pyproject.toml
[tool.poetry]
name = "who-in-school"
version = "0.1.0"
description = ""
authors = ["lvtomatoj <lvtomatoj@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
sqlmodel = "^0.0.16"
fastapi = "^0.110.1"
python-jose = "^3.3.0"
passlib = "^1.7.4"
uvicorn = "^0.29.0"
pydantic-settings = "^2.2.1"
python-multipart = "^0.0.9"
pydantic-extra-types = "^2.6.0"
phonenumbers = "^8.13.34"
wechatpy = "^1.8.18"
cryptography = "^42.0.7"
pycryptodome = "^3.20.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
